backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
mysql-connector-python==8.4.0
psycopg2-binary==2.9.9
google-generativeai==0.7.2
python-dotenv==1.0.1
pandas
openpyxl
python-docx
pypdf
python-multipart

sqlalchemy
